Replace librdkafka event polling with IO event triggering
This PR has two main aims:

* To add a new method to the consumer, PollContext(), which instead
  of accepting a max duration to poll for, accepts a context object
  and waits until either the context is canceled, or an event arrives.
  Importantly, it cancels the underlying call to rd_kafka_poll()
  cleanly once the context is canceled. This is achieved by using
  librdkafka's IO event triggering.
* To replace all places in the consumer and producer implementations
  that perform a similar kind of polling with usage of IO event
  triggering.

This patch uses the rd_kafka_queue_io_event_enable mechanism in
librdkafka. This causes librdkafka to write a byte to a file descriptor
whenever the number of events available transitions from zero to
nonzero. This enables us to "bridge" the world of librdkafka's internal
event loop, and the event loop maintained by the Golang scheduler, so
that we can effectively use a multiplexing `select{}` to wait for
events.

This change is reasonably invasive and unfortunately required the
shuffling around of a _small_ amount of code. The switch statement in
eventPoll has been factored into a new method processRkevToGoEvent,
which either internally handles an event and returns nil, or returns it.
Also, the explicit forwarding of delivery reports to the main events
channel in the RD_KAFKA_EVENT_DR case has been removed, and instead the
producer sets the main channel as the DR channel on the messages
_private field.