Make dynamic build use system librdkafka headers
cgo adds -I${SRCDIR} to the header search path by default; therefore,
using #include <librdkafka/rdkafka.h> can actually still find the
bundled librdkafka header in preference to the system one even when
`-tags=dynamic` is specified.

This change instead makes the static build set a preprocessor macro
USE_VENDORED_LIBRDKAFKA, which is used in a new header rdkafka_select.h
to either import the bundled one (with a special name, and quotes), or
the system one (with the <> import and the expected name).