Rewrite import paths from gopkgin -> github.com
Nowadays, most people should be using github.com anyway, because it's
the recommended way to do it when using go modules. Unfortunately, I
can't actually make the examples use go modules, because they still need
to build on go1.9.

This is needed because I deleted the librdkafka subpackage, but the
gopkg.in/confluentinc/confluent-kafka-go.v1/kafka package that gopkg.in
imports for us depends on
github.com/confluentinc/confluent-kafka-go/kafka/librdkafka, which no
longer exists in GOPATH now. So, in order to get the examples to build,
rewrite the imports.