bundle-import: parse (well known) pkg-config Requires: dependencies too.
As we'll want to avoid having to rely on pkg-config being installed on the
target system we need to make an educated guess on what libraries a
required pkg-config package links to.
For now we limit that to zlib -> -lz.

This should preferably be taken care of by librdkafka's mklove by providing
the libs in a special pkg-config field.