Implement cancelation of calls to rd_kafka_queue_poll
The guts of this implementation is adding a goroutine which runs in
parallel to any call to rd_kafka_queue_poll in eventPoll/pollLogEvents,
which monitors for context cancelation, and if so, runs
rd_kafka_queue_yield to break out of the call to rd_kafka_queue_poll.

The main structural change to eventPoll is that it no longer accepts a
channel as an argument, and instead always returns only zero or one
events for application processing. This is mostly a case of changing
references of `retval = ...` to `return ...` in the main processing
select statement. However, there is one slightly more substantial
change:

Previously, the producer implemented the `go.delivery.reports` setting
by passing along the producer events channel to eventPoll, and having
eventPoll check the value of `h.fwdDr` when deciding what to do with
delivery reports. However, we can't do this anymore, because we're not
passing a channel to eventPoll.

Instead, we now handle h.fwdDr at produce time by setting the _private
field on the message to contain a reference to the `p.events` channel.
This is how dispatching DR's to an alternate per-message channel already
worked, so it doesn't seem a big change to do this, and it allows the
eventPoll code to unconditionally just dispatch DR's to whatever
_private contains.