Ensure rdkafka main queue events are serviced
The main rdkafka queue is where the offset commit events go; if we do
not service this queue, the events will pile up, using memory.

Also we won't know if offsets have been committed, but zendesk_kafka_go
v2 doesn't even expose an API for that.

"Service" the queue by forwarding it to the consumer queue, which we do
service.