[consumer] Propagate errors for Pause/Resume APIs
This propagates the individual errors for each TopicPartition
that is being Paused/Resumed, into the input slice itself.
An non-nil cumulative error is returned in case one or more of
the TopicPartitions has an error.
This approach is different from other such APIs, which return a
new slice of TopicPartitions, but that would break the API.
This method keeps the existing API to not break existing code, but
allows users to access errors which were not being made available
earlier through the input.