Update go version and add developer instructions (#918)
The go version is updated because 1.11 has been long out of support, and parts of the code are already using features only available in 1.13+.

The oldest version of go officially supported (by the golang team) is 1.18, but that is a bit too aggressive for us, and so we go for 1.14.

And as for preferring go1.18 for development, it's because go1.19 has a bunch of changes in gofmt etc, and the release has been out only recently. So we will stick to 1.18 till it's out of support, and then make the changes to all files with the new gofmt.