godoc: Change const groups so enum values are grouped with type
Use the `const Name Type = expression` syntax for all typed enum
constants. This causes the godoc tool to group the consts with the
type definition, which makes it easier to find in the package API
documentation.

The godoc tool does not group consts if the type comes from the right
hand side, as `const Name = Type(expression)`. This is the syntax the
kafka package uses. This means most enum const values are not easily
found. For example, type OffsetSpec does not show the defined
constant values:

https://pkg.go.dev/github.com/confluentinc/confluent-kafka-go/v2@v2.3.0/kafka#OffsetSpec

Instead, the constants are in the package Constants section at the
top:

https://pkg.go.dev/github.com/confluentinc/confluent-kafka-go/v2@v2.3.0/kafka#pkg-constants

The ErrorCode type uses the correct syntax:

https://pkg.go.dev/github.com/confluentinc/confluent-kafka-go/v2@v2.3.0/kafka#ErrorCode

This is much easier to understand. This change removes a redundant
type from the generated ErrorCode syntax since it is unneeded. This
makes the syntax consistent with the other const groups.

For more details, see Go's documentation about const grouping:
https://tip.golang.org/doc/comment#const